Updated Cloning!
Now components can be cloned and maintain their Hierarchy!
e.g. if we clone a child of a component, the cloned child will be attached to the Parent
if we clone a whole hierarchy, that whole hierarchy will be cloned as well as-is!
if we clone the root, it will create  new actor and all the cloned children will come along with it